+++ avatar/providers.py
@@ -24,7 +24,7 @@ class DefaultAvatarProvider(object):
     """
 
     @classmethod
-    def get_avatar_url(cls, user, size):
+    def get_avatar_url(cls, user, width, height):
         return get_default_avatar_url()
 
 
@@ -34,10 +34,10 @@ class PrimaryAvatarProvider(object):
     """
 
     @classmethod
-    def get_avatar_url(cls, user, size):
-        avatar = get_primary_avatar(user, size)
+    def get_avatar_url(cls, user, width, height):
+        avatar = get_primary_avatar(user, width, height)
         if avatar:
-            return avatar.avatar_url(size)
+            return avatar.avatar_url(width, height)
 
 
 class GravatarAvatarProvider(object):
@@ -46,8 +46,8 @@ class GravatarAvatarProvider(object):
     """
 
     @classmethod
-    def get_avatar_url(cls, user, size):
-        params = {"s": str(size)}
+    def get_avatar_url(cls, user, width, _height):
+        params = {"s": str(width)}
         if settings.AVATAR_GRAVATAR_DEFAULT:
             params["d"] = settings.AVATAR_GRAVATAR_DEFAULT
         if settings.AVATAR_GRAVATAR_FORCEDEFAULT:
@@ -68,11 +68,11 @@ class FacebookAvatarProvider(object):
     """
 
     @classmethod
-    def get_avatar_url(cls, user, size):
+    def get_avatar_url(cls, user, width, height):
         fb_id = get_facebook_id(user)
         if fb_id:
-            url = "https://graph.facebook.com/{fb_id}/picture?type=square&width={size}&height={size}"
-            return url.format(fb_id=fb_id, size=size)
+            url = "https://graph.facebook.com/{fb_id}/picture?type=square&width={width}&height={height}"
+            return url.format(fb_id=fb_id, width=width, height=height)
 
 
 class InitialsAvatarProvider(object):
@@ -82,13 +82,13 @@ class InitialsAvatarProvider(object):
     """
 
     @classmethod
-    def get_avatar_url(cls, user, size):
+    def get_avatar_url(cls, user, width, _height):
         initials = user.first_name[:1] + user.last_name[:1]
         if not initials:
             initials = user.username[:1]
         initials = initials.upper()
         context = {
-            "fontsize": (size * 1.1) / 2,
+            "fontsize": (width * 1.1) / 2,
             "initials": initials,
             "hue": user.pk % 360,
             "saturation": "65%",