+++ avatar/urls.py
@@ -7,7 +7,12 @@ urlpatterns = [
     re_path(r"^change/$", views.change, name="avatar_change"),
     re_path(r"^delete/$", views.delete, name="avatar_delete"),
     re_path(
-        r"^render_primary/(?P<user>[\w\d\@\.\-_]+)/(?P<size>[\d]+)/$",
+        r"^render_primary/(?P<user>[\w\d\@\.\-_]+)/(?P<width>[\d]+)/$",
+        views.render_primary,
+        name="avatar_render_primary",
+    ),
+    re_path(
+        r"^render_primary/(?P<user>[\w\d\@\.\-_]+)/(?P<width>[\d]+)/(?P<height>[\d]+)/$",
         views.render_primary,
         name="avatar_render_primary",
     ),