+++ avatar/utils.py
@@ -18,18 +18,23 @@ def get_username(user):
         return user.username
 
 
-def get_user(username):
-    """Return user from a username/ish identifier"""
-    return get_user_model().objects.get_by_natural_key(username)
+def get_user(userdescriptor):
+    """Return user from a username/ID/ish identifier"""
+    User = get_user_model()
+    if userdescriptor.isdigit():
+        user = User.objects.filter(id=int(userdescriptor)).first()
+        if user:
+            return user
+    return User.objects.get_by_natural_key(userdescriptor)
 
 
-def get_cache_key(user_or_username, size, prefix):
+def get_cache_key(user_or_username, width, height, prefix):
     """
     Returns a cache key consisten of a username and image size.
     """
     if isinstance(user_or_username, get_user_model()):
         user_or_username = get_username(user_or_username)
-    key = "%s_%s_%s" % (prefix, user_or_username, size)
+    key = "%s_%s_%s_%s" % (prefix, user_or_username, width, height or width)
     return "%s_%s" % (
         slugify(key)[:100],
         hashlib.md5(force_bytes(key)).hexdigest(),
@@ -43,8 +48,8 @@ def cache_set(key, value):
 
 def cache_result(default_size=settings.AVATAR_DEFAULT_SIZE):
     """
-    Decorator to cache the result of functions that take a ``user`` and a
-    ``size`` value.
+    Decorator to cache the result of functions that take a ``user``, a
+    ``width`` and a ``height`` value.
     """
     if not settings.AVATAR_CACHE_ENABLED:
 
@@ -54,13 +59,13 @@ def cache_result(default_size=settings.AVATAR_DEFAULT_SIZE):
         return decorator
 
     def decorator(func):
-        def cached_func(user, size=None, **kwargs):
+        def cached_func(user, width=None, height=None, **kwargs):
             prefix = func.__name__
             cached_funcs.add(prefix)
-            key = get_cache_key(user, size or default_size, prefix=prefix)
+            key = get_cache_key(user, width or default_size, height, prefix=prefix)
             result = cache.get(key)
             if result is None:
-                result = func(user, size or default_size, **kwargs)
+                result = func(user, width or default_size, height, **kwargs)
                 cache_set(key, result)
             return result
 
@@ -69,16 +74,20 @@ def cache_result(default_size=settings.AVATAR_DEFAULT_SIZE):
     return decorator
 
 
-def invalidate_cache(user, size=None):
+def invalidate_cache(user, width=None, height=None):
     """
-    Function to be called when saving or changing an user's avatars.
+    Function to be called when saving or changing a user's avatars.
     """
     sizes = set(settings.AVATAR_AUTO_GENERATE_SIZES)
-    if size is not None:
-        sizes.add(size)
+    if width is not None:
+        sizes.add((width, height or width))
     for prefix in cached_funcs:
         for size in sizes:
-            cache.delete(get_cache_key(user, size, prefix))
+            if isinstance(size, int):
+                cache.delete(get_cache_key(user, size, size, prefix))
+            else:
+                # Size is specified with height and width.
+                cache.delete(get_cache_key(user, size[0], size[1], prefix))
 
 
 def get_default_avatar_url():
@@ -101,7 +110,7 @@ def get_default_avatar_url():
     return "%s%s" % (base_url, settings.AVATAR_DEFAULT_URL)
 
 
-def get_primary_avatar(user, size=settings.AVATAR_DEFAULT_SIZE):
+def get_primary_avatar(user, width=settings.AVATAR_DEFAULT_SIZE, height=None):
     User = get_user_model()
     if not isinstance(user, User):
         try:
@@ -117,6 +126,6 @@ def get_primary_avatar(user, size=settings.AVATAR_DEFAULT_SIZE):
     except IndexError:
         avatar = None
     if avatar:
-        if not avatar.thumbnail_exists(size):
-            avatar.create_thumbnail(size)
+        if not avatar.thumbnail_exists(width, height):
+            avatar.create_thumbnail(width, height)
     return avatar